refactor(snapshots): collapse two endpoints into one to fit Hobby cap
Vercel Hobby caps a deployment at 12 serverless functions. The repo was
already at 12 before this branch, so the two snapshot files I added
(api/snapshots/index.js + api/snapshots/[id].js) pushed it to 14 and
broke deploy.

Consolidates both into a single api/snapshots.js that dispatches by
method + optional ?id= query param. Net function count: 13. Still 1
over the Hobby cap; further consolidation needed elsewhere.

  POST   /api/snapshots         -> save
  GET    /api/snapshots         -> list
  GET    /api/snapshots?id=...  -> load one
  DELETE /api/snapshots?id=...  -> delete one

Client (snapshotClient.ts) updated to use ?id= instead of path-style
URLs. No behavior change.
