+++ api/snapshots.js
@@ -1,18 +1,28 @@
 import crypto from 'crypto';
-import { put, list } from '@vercel/blob';
-import { json, methodNotAllowed } from '../_lib/response.js';
-import { enforceRateLimit } from '../_lib/rateLimit.js';
-import { requireOwner } from '../_lib/ownerAuth.js';
+import { put, list, del } from '@vercel/blob';
+import { json, methodNotAllowed } from './_lib/response.js';
+import { enforceRateLimit } from './_lib/rateLimit.js';
+import { requireOwner } from './_lib/ownerAuth.js';
+
+// Single endpoint for the whole snapshot lifecycle. Vercel Hobby caps a
+// deployment at 12 serverless functions, so we dispatch by method + the
+// optional `?id=<uuid>` query param instead of splitting into per-id and
+// collection files.
+//
+//   POST   /api/snapshots         -> save (body: { title, project, images })
+//   GET    /api/snapshots         -> list summaries
+//   GET    /api/snapshots?id=...  -> load one
+//   DELETE /api/snapshots?id=...  -> remove one
 
 // Hard cap on a single snapshot payload — guards against runaway uploads
-// (e.g. accidentally bundling a giant project). 60 MB comfortably fits a
-// project with ~20 high-quality gpt-image-2 PNGs (~1-3 MB each as base64).
-const MAX_BODY_BYTES = 60 * 1024 * 1024;
+// (e.g. accidentally bundling a giant project). Vercel's Hobby request
+// body cap is ~4.5 MB; this matches.
+const MAX_BODY_BYTES = 4 * 1024 * 1024;
 
 const SNAPSHOT_PREFIX = 'snapshots/';
+const ID_RE = /^[0-9a-f-]{8,64}$/i;
 
 async function readJsonBody(req) {
-  // Vercel may pre-parse JSON. If req.body is a string/object already, prefer it.
   if (req.body && typeof req.body === 'object') return req.body;
   if (typeof req.body === 'string' && req.body.length > 0) {
     return JSON.parse(req.body);
@@ -41,14 +51,23 @@ async function readJsonBody(req) {
   });
 }
 
-function nowIso() {
-  return new Date().toISOString();
-}
+const nowIso = () => new Date().toISOString();
 
 function sanitizeTitle(title) {
   if (typeof title !== 'string') return 'Untitled';
   const trimmed = title.trim().slice(0, 200);
   return trimmed.length > 0 ? trimmed : 'Untitled';
+}
+
+async function findBlobsForId(id) {
+  const out = [];
+  let cursor;
+  do {
+    const page = await list({ prefix: `${SNAPSHOT_PREFIX}${id}/`, cursor });
+    for (const blob of page.blobs) out.push(blob);
+    cursor = page.hasMore ? page.cursor : undefined;
+  } while (cursor);
+  return out;
 }
 
 async function handlePost(req, res) {
@@ -78,18 +97,10 @@ async function handlePost(req, res) {
     imageCount: images.length,
   };
 
-  const data = {
-    schemaVersion: 1,
-    manifest,
-    project,
-    images,
-  };
-
+  const data = { schemaVersion: 1, manifest, project, images };
   const dataJson = JSON.stringify(data);
   const manifestJson = JSON.stringify({ ...manifest, sizeBytes: dataJson.length });
 
-  // Sequential writes: manifest second so a partial failure leaves an orphaned
-  // data blob (cleanable) rather than a manifest pointing at nothing.
   await put(`${SNAPSHOT_PREFIX}${id}/data.json`, dataJson, {
     contentType: 'application/json',
     access: 'public',
@@ -107,8 +118,6 @@ async function handlePost(req, res) {
 }
 
 async function handleList(_req, res) {
-  // Vercel Blob `list` returns up to 1000 entries per page. For a personal
-  // portfolio that's effectively unbounded. We page through to be safe.
   const summaries = [];
   let cursor;
   do {
@@ -118,8 +127,7 @@ async function handleList(_req, res) {
       try {
         const resp = await fetch(blob.url, { cache: 'no-store' });
         if (!resp.ok) continue;
-        const manifest = await resp.json();
-        summaries.push(manifest);
+        summaries.push(await resp.json());
       } catch {
         // skip unreadable manifests
       }
@@ -131,13 +139,30 @@ async function handleList(_req, res) {
   return json(res, 200, { snapshots: summaries });
 }
 
+async function handleGetOne(id, res) {
+  const blobs = await findBlobsForId(id);
+  const dataBlob = blobs.find((b) => b.pathname.endsWith('/data.json'));
+  if (!dataBlob) return json(res, 404, { error: 'not_found' });
+
+  const resp = await fetch(dataBlob.url, { cache: 'no-store' });
+  if (!resp.ok) return json(res, 502, { error: 'blob_fetch_failed' });
+  return json(res, 200, await resp.json());
+}
+
+async function handleDelete(id, res) {
+  const blobs = await findBlobsForId(id);
+  if (blobs.length === 0) return json(res, 404, { error: 'not_found' });
+  await Promise.all(blobs.map((b) => del(b.url)));
+  return json(res, 200, { id, deleted: blobs.length });
+}
+
 export default async function handler(req, res) {
-  if (req.method !== 'POST' && req.method !== 'GET') {
-    return methodNotAllowed(res, ['GET', 'POST']);
+  if (!['GET', 'POST', 'DELETE'].includes(req.method)) {
+    return methodNotAllowed(res, ['GET', 'POST', 'DELETE']);
   }
   if (
     enforceRateLimit(req, res, {
-      scope: 'snapshots_index',
+      scope: 'snapshots',
       limit: 60,
       windowMs: 60_000,
       errorBody: { error: 'rate_limited' },
@@ -147,9 +172,23 @@ export default async function handler(req, res) {
   }
   if (requireOwner(req, res)) return;
 
+  const id = typeof req.query?.id === 'string' ? req.query.id : null;
+  if (id !== null && !ID_RE.test(id)) {
+    return json(res, 400, { error: 'invalid_id' });
+  }
+
   try {
-    if (req.method === 'POST') return await handlePost(req, res);
-    return await handleList(req, res);
+    if (req.method === 'POST') {
+      if (id !== null) return json(res, 400, { error: 'unexpected_id' });
+      return await handlePost(req, res);
+    }
+    if (req.method === 'DELETE') {
+      if (id === null) return json(res, 400, { error: 'missing_id' });
+      return await handleDelete(id, res);
+    }
+    // GET
+    if (id === null) return await handleList(req, res);
+    return await handleGetOne(id, res);
   } catch (err) {
     console.error('[snapshots]', err);
     return json(res, 500, { error: 'internal_error', message: err?.message ?? 'unknown' });
